refactor - resolved circular dependency by moving calls to third class
initially reviewsDatabaseFunctions would call updateOverallRating for a tutor after adding a review. This meant adding an instance of tutorDatabaseFunctions to the reviewsDatabaseFunctions class. However, updateOverallReview needed an instance of reviewDatabaseFunctions since it needed to get all of a tutor's reviews before their overall rating could be updated.
This problem was resolved by moving both function calls to reviewsManager class after the its call to addReview.